move uitools to end of supervisor
