+++ app/services/agent/__init__.py
@@ -1,3 +1,2 @@
 from .child import create_child_agent
 from .supervisor import create_supervisor_agent
-from .state import AgentState
